pyproject.toml
[project]
name = "telegram-formatter"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.115.12",
    "markdown-it-py==4.0.0",
    "uvicorn==0.34.0",
    "pydantic-settings==2.9.1",
]

########################################
# 3) dev-зависимости
########################################
[project.optional-dependencies]
dev = [
  "pytest-asyncio==1.0.0",
  "httpx==0.28.1",
  "mypy==1.16.0",
  "asgi-lifespan==2.1.0",
  "ruff",
]

########################################
# 4) Где искать пакеты (каталог app/)
########################################
[tool.setuptools.packages.find]
where = [""]
exclude = ["tests*"]

########################################
# 7) Линтеры / форматтеры
########################################
[tool.ruff]
line-length    = 120
target-version = "py312"
src            = ["app"]

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.lint.isort]
section-order = [
    "future",
    "standard-library",
    "third-party",
    "first-party",
    "local-folder"
]
lines-after-imports = 2
force-sort-within-sections = true
combine-as-imports = false
